#1 Done! There's correct field. But maybe needs some refactoring
